+++ CHANGELOG.md
@@ -55,6 +55,7 @@ Senior Web Developer
 - **(enh)** Added CONTAINER_PREFIX constant for html div container id
 - **(fix)** Added class "hide" to modal div
 - **(fix)** Add support for non-link brand in TbNavbar
+- **(enh)** Update x-editable-yii from 1.3.0 to 1.4.0 #413 (magefad)
 
 
 ### YiiBooster version 1.0.5